Upgrade dependencies and more code cleanup
