+++ .eslintrc.js
@@ -9,6 +9,5 @@ module.exports = {
   ],
   rules: {
     "mozilla/no-define-cc-etc": "off", // seems broken outside of m-c
-    "react/prop-types": "off", // TODO
   },
 };
