--- extension/.eslintrc
@@ -1,3 +1,0 @@
-{
-  
-}
