+++ extension/content/components/filters/JexlColumn.js
@@ -1,3 +1,4 @@
+import PropTypes from "prop-types";
 import React from "react";
 import { Controlled as CodeMirror } from "react-codemirror2";
 
@@ -7,6 +8,11 @@ import "codemirror/mode/javascript/javascript";
 import "codemirror/theme/neo.css";
 
 export default class JexlColumn extends React.PureComponent {
+  static propTypes = {
+    filterExpression: PropTypes.object.isRequired,
+    onBeforeChange: PropTypes.func.isRequired,
+  };
+
   render() {
     const { filterExpression, onBeforeChange } = this.props;
     return (
