+++ extension/content/less/filter-page.less
@@ -8,9 +8,9 @@
     height: 40px;
     left: 0;
     line-height: 20px;
+    overflow: hidden;
     padding: 10px;
     position: absolute;
-    overflow: hidden;
     text-overflow: ellipsis;
     top: 0;
     white-space: nowrap;
@@ -52,12 +52,12 @@
 
   .rc-tree {
     padding: 0 10px;
-    user-select: none;
     -moz-user-select: none;
+    user-select: none;
 
     li span.rc-tree-switcher:not(.rc-tree-switcher-noop) {
-      text-align: center;
       background: none;
+      text-align: center;
     }
 
     .rc-tree-title {
