+++ extension/content/less/index.less
@@ -1,9 +1,9 @@
 @import "~rsuite/styles/less/index.less";
 @import "~highlight.js/styles/nord.css";
 @import '~codemirror/lib/codemirror.css';
-
+// Theme
 @import "theme";
-
+// Extension styling
 @import "main";
 @import "split";
 @import "filter-page";
