+++ extension/content/less/main.less
@@ -9,8 +9,8 @@
 .app-sidebar {
   background: @B050;
   height: 100vh;
-  overflow: auto;
   left: 0;
+  overflow: auto;
   position: fixed;
   top: 0;
 
@@ -22,8 +22,8 @@
     position: relative;
 
     svg {
-      margin-top: -15px;
       left: 14px;
+      margin-top: -15px;
       position: absolute;
       top: 50%;
       width: 30px;
