+++ extension/content/less/split.less
@@ -17,11 +17,11 @@
       border-right: 1px dashed @blue;
     }
 
-    &:before {
+    &::before {
       content: '';
       height: 100%;
-      position: absolute;
       left: -5px;
+      position: absolute;
       top: 0;
       width: 11px;
     }
