+++ package.json
@@ -20,14 +20,14 @@
     "watch:webpack": "webpack --progress --mode development --watch"
   },
   "devDependencies": {
-    "babel-core": "<7",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-async-generator-functions": "^7.2.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
     "babel-eslint": "^8.2.3",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-import": "^1.7.0",
-    "babel-plugin-transform-async-generator-functions": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.5",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-react": "^6.24.1",
+    "babel-loader": "^8.0.6",
     "copy-webpack-plugin": "^4.5.1",
     "css-loader": "^0.28.11",
     "eslint": "^5.16.0",
@@ -45,13 +45,15 @@
     "npm-run-all": "^4.1.3",
     "prettier": "^1.12.1",
     "style-loader": "^0.21.0",
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-order": "^3.0.0",
     "web-ext": "^2.6.0",
     "webpack": "^4.8.3",
     "webpack-cli": "^2.1.3"
   },
   "dependencies": {
     "autobind-decorator": "^2.4.0",
-    "autoscroll-react": "^3.2.0",
     "codemirror": "^5.47.0",
     "js-yaml": "^3.11.0",
     "luxon": "^1.2.1",
